# perf: neon Paeth unfiltering (3bpp)